Implementación módulo auditoría y recuperación de contraseña: vista de logs de auditoría en Angular con filtros por usuario, tipo de evento y rango de fechas; tabla dinámica con paginación; botón de limpiar filtros; actualización en tiempo real.

+++ src/main/java/com/uniminuto/clinica/model/RecetaRq.java
@@ -13,19 +13,19 @@ public class RecetaRq {
     /**
      * Id de la receta.
      */
-    private Integer id;
+    private Long id;
 
     /**
      * Identificador de la cita asociada a la receta.
      */
     @NotNull(message = "El identificador de la cita es obligatorio.")
-    private Integer citaId;
+    private Long citaId;
 
     /**
      * Identificador del medicamento recetado.
      */
     @NotNull(message = "El identificador del medicamento es obligatorio.")
-    private Integer medicamentoId;
+    private Long medicamentoId;
 
     /**
      * Dosis prescrita del medicamento.
